LPS-197423 Migrate layoutDataReducer to typescript
