+++ modules/apps/layout/layout-content-page-editor-web/types/src/main/resources/META-INF/resources/page_editor/app/reducers/layoutDataReducer.d.ts
@@ -6,7 +6,9 @@
 import {LayoutData} from '../../types/layout_data/LayoutData';
 import addFragmentEntryLinks from '../actions/addFragmentEntryLinks';
 import addItem from '../actions/addItem';
+import addRule from '../actions/addRule';
 import deleteItem from '../actions/deleteItem';
+import deleteRule from '../actions/deleteRule';
 import duplicateItem from '../actions/duplicateItem';
 import moveItem from '../actions/moveItem';
 import updateColSize from '../actions/updateColSize';
@@ -17,21 +19,25 @@ import updateItemConfig from '../actions/updateItemConfig';
 import updateItemLocalConfig from '../actions/updateItemLocalConfig';
 import updatePreviewImage from '../actions/updatePreviewImage';
 import updateRowColumns from '../actions/updateRowColumns';
+import updateRule from '../actions/updateRule';
 export declare const INITIAL_STATE: LayoutData;
 export default function layoutDataReducer(
 	layoutData: LayoutData | undefined,
 	action:
-		| ReturnType<typeof updateColSize>
-		| ReturnType<typeof updateCollectionDisplayCollection>
 		| ReturnType<typeof addFragmentEntryLinks>
 		| ReturnType<typeof addItem>
+		| ReturnType<typeof addRule>
 		| ReturnType<typeof deleteItem>
+		| ReturnType<typeof deleteRule>
 		| ReturnType<typeof duplicateItem>
 		| ReturnType<typeof moveItem>
-		| ReturnType<typeof updateFragmentEntryLinkConfiguration>
-		| ReturnType<typeof updateRowColumns>
+		| ReturnType<typeof updateCollectionDisplayCollection>
+		| ReturnType<typeof updateColSize>
 		| ReturnType<typeof updateFormItemConfig>
+		| ReturnType<typeof updateFragmentEntryLinkConfiguration>
 		| ReturnType<typeof updateItemConfig>
 		| ReturnType<typeof updateItemLocalConfig>
 		| ReturnType<typeof updatePreviewImage>
+		| ReturnType<typeof updateRowColumns>
+		| ReturnType<typeof updateRule>
 ): LayoutData;
